New One Page Mockups & Assets
Made layouts for one page test app with scoreboard on top and control panel layout below. Also made mockup of what it looks like as video is selected and lightbox modal layer opens and the selected video is playing on the stage (and lightbox controls). Also made png images used to activate controls to close lightbox and pop-out and pop-back in from full screen view for scoreboard or control panel section/divs.